Add files via upload
Save Size button: set the user setting for the size to the current client size and save the settings. 
Save Location button: set the user setting for the location to the current desktop location and save the settings. 
Reset Settings button: reset the user settings to the original default values. 
    Set the client size and desktop location to the reset settings.