Mohammed module 2 (#10)
* Add files via upload

Got the Message box to show up on closing

* Add files via upload

Got the dialog box to cancel closing the application

* Add files via upload

Got the dialog box to cancel closing the application

* Add files via upload

Finished up my part of the project. Message box shows up on closing to cancel and the application hides when it loses focus.