Add files via upload
Made the OathDialog re-sizable and used the path gradient brush do design a background for the form. In the AboutDialog, I made it re-sizable also, changed the description to fit this module 6, and started working on making the pen for the borders (It is currently filling the entire background, but it was to check the pen was working).